+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/config/TeaBuildConfiguration.java
@@ -16,9 +16,6 @@ public class TeaBuildConfiguration {
     public ArrayList<File> assetsPath = new ArrayList<>();
     public ArrayList<String> additionalAssetsClasspathFiles = new ArrayList<>();
 
-    @Deprecated
-    public boolean obfuscate = false;
-
     private String mainApplicationClass;
 
     public String webappPath;
@@ -38,10 +35,15 @@ public class TeaBuildConfiguration {
     public String htmlTitle = "gdx-teavm";
     public int htmlWidth = 800;
     public int htmlHeight = 600;
-    /**
-     * If the logo is shown while the application is loading.
-     */
+
+    /** True to use the default html index. False will stop overwriting html file. */
+    public boolean useDefaultHtmlIndex = true;
+
+    /** If the logo is shown while the application is loading. Requires showLoadingLogo true. */
     public boolean showLoadingLogo = true;
+
+    /** Logo asset path. Requires showLoadingLogo true. */
+    public String logoPath = "startup-logo.png";
 
     public String getHtmlTitle() {
         return htmlTitle;
@@ -57,6 +59,10 @@ public class TeaBuildConfiguration {
 
     public boolean isShowLoadingLogo() {
       return showLoadingLogo;
+    }
+
+    public boolean shouldUseDefaultHtmlIndex() {
+      return useDefaultHtmlIndex;
     }
 
     public String getMainClass() {
@@ -92,11 +98,6 @@ public class TeaBuildConfiguration {
         return assetFilter;
     }
 
-    @Deprecated
-    public boolean minifying() {
-        return obfuscate;
-    }
-
     public ArrayList<String> getReflectionInclude() {
         return reflectionInclude;
     }
@@ -126,5 +127,9 @@ public class TeaBuildConfiguration {
 
     public boolean acceptClasspath(URL url) {
         return true;
+    }
+
+    public String getLogoPath() {
+        return logoPath;
     }
 }