+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/config/TeaBuilder.java
@@ -472,7 +472,6 @@ public class TeaBuilder {
 
         File setTargetDirectory = new File(webappDirectory + File.separator + webappName + File.separator + "teavm");
         String setTargetFileName = "app.js";
-        boolean setMinifying = configuration.minifying();
         String tmpdir = System.getProperty("java.io.tmpdir");
         File setCacheDirectory = new File(tmpdir + File.separator + "TeaVMCache");
         boolean setIncremental = false;
@@ -483,7 +482,6 @@ public class TeaBuilder {
         tool.setSourceFilesCopied(setSourceFilesCopied);
         tool.setTargetDirectory(setTargetDirectory);
         tool.setTargetFileName(setTargetFileName);
-        tool.setObfuscated(setMinifying);
         tool.setFastDependencyAnalysis(false);
         tool.setOptimizationLevel(TeaVMOptimizationLevel.SIMPLE);
         String applicationListenerClass = configuration.getApplicationListenerClass();
@@ -535,35 +533,42 @@ public class TeaBuilder {
         webappAssetsFiles.add(webappName);
         TeaBuilder.logHeader("COPYING ASSETS");
 
-        AssetsCopy.copy(classLoader, webappAssetsFiles, new ArrayList<>(), null, webappDirectory, false);
-        TeaBuilder.log("");
+        boolean shouldUseDefaultHtmlIndex = configuration.shouldUseDefaultHtmlIndex();
+        if(shouldUseDefaultHtmlIndex) {
+            AssetsCopy.copy(classLoader, webappAssetsFiles, new ArrayList<>(), null, webappDirectory, false);
+            TeaBuilder.log("");
+        }
 
         String scriptsOutputPath = webappDirectory + File.separator + webappName;
         String assetsOutputPath = scriptsOutputPath + File.separator + "assets";
 
-        File indexFile = new File(scriptsOutputPath + File.separator + "index.html");
-        FileHandle handler = new FileHandle(indexFile);
-        String indexHtmlStr = handler.readString();
-
-        String logo = "assets/startup-logo.png";
-        indexHtmlStr = indexHtmlStr.replace("%TITLE%", configuration.getHtmlTitle());
-        indexHtmlStr = indexHtmlStr.replace("%WIDTH%", configuration.getHtmlWidth());
-        indexHtmlStr = indexHtmlStr.replace("%HEIGHT%", configuration.getHtmlHeight());
-        indexHtmlStr = indexHtmlStr.replace("%ARGS%", configuration.getMainClassArgs());
-        indexHtmlStr = indexHtmlStr.replace("%LOGO%",
-            configuration.isShowLoadingLogo() ? "<img id=\"progress-img\" src=\"" + logo + "\">"
-                                              : ""
-        );
-
-        handler.writeString(indexHtmlStr, false);
-
         AssetFilter filter = configuration.assetFilter();
         ArrayList<File> assetsPaths = new ArrayList<>();
         ArrayList<String> classPathAssetsFiles = new ArrayList<>();
-        ArrayList<String> classPathScriptFiles = new ArrayList<>();
         assetsDefaultClasspath(classPathAssetsFiles);
-        if (configuration.isShowLoadingLogo()) {
-          classPathScriptFiles.add(logo);
+
+        if(shouldUseDefaultHtmlIndex) {
+            File indexFile = new File(scriptsOutputPath + File.separator + "index.html");
+            FileHandle handler = new FileHandle(indexFile);
+            String indexHtmlStr = handler.readString();
+
+            String logo = configuration.getLogoPath();
+            String htmlLogo = "assets/" + logo;
+            boolean showLoadingLogo = configuration.isShowLoadingLogo();
+
+            indexHtmlStr = indexHtmlStr.replace("%TITLE%", configuration.getHtmlTitle());
+            indexHtmlStr = indexHtmlStr.replace("%WIDTH%", configuration.getHtmlWidth());
+            indexHtmlStr = indexHtmlStr.replace("%HEIGHT%", configuration.getHtmlHeight());
+            indexHtmlStr = indexHtmlStr.replace("%ARGS%", configuration.getMainClassArgs());
+            indexHtmlStr = indexHtmlStr.replace(
+                    "%LOGO%", showLoadingLogo ? "<img id=\"progress-img\" src=\"" + htmlLogo + "\">" : ""
+            );
+
+            handler.writeString(indexHtmlStr, false);
+
+            if(showLoadingLogo) {
+                classPathAssetsFiles.add(logo);
+            }
         }
         ArrayList<String> additionalAssetClasspath = configuration.getAdditionalAssetClasspath();
         classPathAssetsFiles.addAll(additionalAssetClasspath);