+++ examples/core/teavm/src/main/java/com/github/xpenatan/gdx/examples/teavm/BuildReflectionTest.java
@@ -19,7 +19,6 @@ public class BuildReflectionTest {
         TeaBuildConfiguration teaBuildConfiguration = new TeaBuildConfiguration();
         teaBuildConfiguration.assetsPath.add(new File("../desktop/assets"));
         teaBuildConfiguration.webappPath = new File("build/dist").getCanonicalPath();
-        teaBuildConfiguration.obfuscate = false;
         teaBuildConfiguration.setApplicationListener(ReflectionTest.class);
         TeaVMTool tool = TeaBuilder.config(teaBuildConfiguration);
         TeaBuilder.build(tool);