+++ examples/core/teavm/src/main/java/com/github/xpenatan/gdx/examples/teavm/BuildTeaVMTestDemo.java
@@ -3,21 +3,22 @@ package com.github.xpenatan.gdx.examples.teavm;
 import com.github.xpenatan.gdx.backends.teavm.config.TeaBuildConfiguration;
 import com.github.xpenatan.gdx.backends.teavm.config.TeaBuilder;
 import com.github.xpenatan.gdx.backends.teavm.gen.SkipClass;
-import com.github.xpenatan.gdx.examples.teavm.launcher.GearsTestLauncher;
+import com.github.xpenatan.gdx.examples.teavm.launcher.TeaVMTestLauncher;
 import java.io.File;
 import java.io.IOException;
 import org.teavm.tooling.TeaVMTool;
 
 @SkipClass
-public class BuildGearsDemo {
+public class BuildTeaVMTestDemo {
 
     public static void main(String[] args) throws IOException {
         TeaBuildConfiguration teaBuildConfiguration = new TeaBuildConfiguration();
         teaBuildConfiguration.assetsPath.add(new File("../desktop/assets"));
         teaBuildConfiguration.webappPath = new File("build/dist").getCanonicalPath();
+        teaBuildConfiguration.logoPath = "logo.png";
 
         TeaVMTool tool = TeaBuilder.config(teaBuildConfiguration);
-        tool.setMainClass(GearsTestLauncher.class.getName());
+        tool.setMainClass(TeaVMTestLauncher.class.getName());
         tool.setObfuscated(false);
         TeaBuilder.build(tool);
     }