+++ examples/core/teavm/src/main/java/com/github/xpenatan/gdx/examples/teavm/launcher/TeaVMTestLauncher.java
@@ -2,16 +2,16 @@ package com.github.xpenatan.gdx.examples.teavm.launcher;
 
 import com.github.xpenatan.gdx.backends.teavm.TeaApplication;
 import com.github.xpenatan.gdx.backends.teavm.TeaApplicationConfiguration;
-import com.github.xpenatan.gdx.examples.tests.GearsDemo;
-import com.github.xpenatan.gdx.examples.tests.ReadPixelsTest;
+import com.github.xpenatan.gdx.examples.tests.TeaVMInputTest;
 
-public class GearsTestLauncher {
+public class TeaVMTestLauncher {
 
     public static void main(String[] args) {
         TeaApplicationConfiguration config = new TeaApplicationConfiguration("canvas");
         config.width = 0;
         config.height = 0;
 //        new TeaApplication(new GearsDemo(), config);
-        new TeaApplication(new ReadPixelsTest(), config);
+//        new TeaApplication(new ReadPixelsTest(), config);
+        new TeaApplication(new TeaVMInputTest(), config);
     }
 }