+++ examples/gdx-tests/core/src/main/java/com/github/xpenatan/imgui/example/tests/imgui/ImGuiGame.java
@@ -1,0 +1,11 @@
+package com.github.xpenatan.imgui.example.tests.imgui;
+
+import com.badlogic.gdx.Game;
+
+public class ImGuiGame extends Game {
+
+    @Override
+    public void create() {
+        setScreen(new ImGuiInitScreen(this));
+    }
+}