+++ examples/gdx-tests/teavm/src/main/java/com/github/xpenatan/gdx/examples/teavm/launcher/GdxTestLauncher.java
@@ -10,6 +10,8 @@ import com.badlogic.gdx.tests.gles3.GL30Texture3DTest;
 import com.badlogic.gdx.tests.gles3.InstancedRenderingTest;
 import com.github.xpenatan.gdx.backends.teavm.TeaApplicationConfiguration;
 import com.github.xpenatan.gdx.backends.teavm.TeaApplication;
+import com.github.xpenatan.imgui.example.tests.imgui.ImGuiGame;
+import com.github.xpenatan.imgui.example.tests.wrapper.TeaVMTestWrapper;
 
 public class GdxTestLauncher {
 
@@ -20,15 +22,7 @@ public class GdxTestLauncher {
         config.showDownloadLogs = true;
         config.useGL30 = true;
         config.useGLArrayBuffer = true;
-//        new TeaApplication(new ImGuiTestsApp(), config);
-//        new TeaApplication(new TeaVMTestWrapper(), config);
-//        new TeaApplication(new SoundTest(), config);
-//        new TeaApplication(new TextureDataTest(), config);
-        new TeaApplication(new FloatTextureTest(), config);
-//        new TeaApplication(new TextureRegion3DTest(), config);
-//        new TeaApplication(new GL30Texture3DTest(), config);
-//        new TeaApplication(new InstancedRenderingTest(), config);
-//        new TeaApplication(new TextureArrayTest(), config);
-//        new TeaApplication(new MultipleRenderTargetTest(), config);
+//        new TeaApplication(new ImGuiGame(), config);
+        new TeaApplication(new TeaVMTestWrapper(), config);
     }
 }