+++ tools/generator/desktop/src/main/java/com/github/xpenatan/gdx/html5/generator/Main.java
@@ -2,7 +2,6 @@ package com.github.xpenatan.gdx.html5.generator;
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.github.xpenatan.teavm.generator.core.view.MainApplication;
 
 public class Main {
     public static void main(String[] args) {
@@ -11,6 +10,6 @@ public class Main {
         config.height = 174;
         config.resizable = false;
         config.title = "Gdx-Html5-Generator";
-        new LwjglApplication(new MainApplication(), config);
+        new LwjglApplication(new ImGuiGame(), config);
     }
 }